% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_norway.R
\name{generate_county_colors}
\alias{generate_county_colors}
\title{Function | generate_county_colors}
\usage{
generate_county_colors(names_df, palette_name = "batlow")
}
\arguments{
\item{names_df}{Data frame with county names}

\item{palette_name}{Name of the scico palette (default: "batlow")}
}
\value{
Named vector of colors for each county
}
\description{
Generate color palette for Norway counties
}
