% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{no_county_names_2024}
\alias{no_county_names_2024}
\title{Dataset | Norway County Names and Location Codes (2024 borders)}
\format{
A dataframe with 15 rows and 2 columns:
\describe{
\item{location_code}{County code identifier (e.g., "county03")}
\item{location_name}{County name in Norwegian (e.g., "Oslo", "Trondelag")}
}
}
\source{
\describe{
\item{csdata package}{\url{https://github.com/csids/csdata}}
\item{Border year}{2024}
}
}
\usage{
no_county_names_2024
}
\description{
Names and location codes for Norwegian counties using 2024 administrative borders.
}
\details{
Data size: 10.1 Kb
}
\references{
Norwegian county names and codes from csdata package.
Data accessed on 2025-09-23
}
\keyword{datasets}
