% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_norway.R
\name{plot_norway_counties}
\alias{plot_norway_counties}
\title{Function | plot_norway_counties}
\usage{
plot_norway_counties(
  map_df = no_county_geodata_2024,
  names_df = no_county_names_2024,
  palette_name = "batlow",
  legend_position = c(0.9, 0.2)
)
}
\arguments{
\item{map_df}{Data frame with map data (default: no_county_geodata_2024)}

\item{names_df}{Data frame with county names (default: no_county_names_202}

\item{palette_name}{Name of the scico palette (default: "batlow")}

\item{legend_position}{Position of the legend (default: c(0.9, 0.2))}
}
\value{
ggplot object of the Norway county map
}
\description{
Create Norway county map with customizable styling
}
