% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_norway.R
\name{plot_trondelag}
\alias{plot_trondelag}
\title{Function | plot_trondelag}
\usage{
plot_trondelag(
  geodata_df = no_county_geodata_2024,
  names_df = no_county_names_2024,
  trondelag_color = "#005E5D",
  other_color = "#FCFAF6",
  legend_position = c(0.7, 0.1)
)
}
\arguments{
\item{geodata_df}{Data frame with map data (default: no_county_geodata_2024)}

\item{names_df}{Data frame with county names (default: no_county_names_2024)}

\item{trondelag_color}{Color for Trondelag (default: Sjogronn "#005E5D")}

\item{other_color}{Color for other counties (default: Beige Lys "#FCFAF6")}

\item{legend_position}{Position of the legend (default: c(0.9, 0.2))}
}
\value{
ggplot object of the Norway county map with Trondelag highlighted
}
\description{
Create Norway county map highlighting Trondelag
}
